Add Attacher#assign_remote_url for dynamically passing downloader options
In certain cases one might need to pass downloader options dynamically
in the controller. For example, they might need to pass certain request
headers that are specific to that certain request.

This adds a new Attacher#assign_remote_url method, which accepts
additional options. Attacher#remote_url= is kept as an alias for
Attacher#assign_remote_url, it will probably remain like that
indefinitely (I don't see a reason for deprecating it).