Fix inheritance with dynamic_storage plugin
On subclassing the dynamic storage resolvers are currently not being
inherited from the Shrine superclass. This is because, instead of saving
the resolvers in Shrine.opts, which is inherited on subclassing, we're
saving them in an instance variable which doesn't get transferred to the
subclass.

This commit modifies the dynamic_storage plugin to properly save
storage resolvers into Shrine.opts.