Better unique ids for multiple file upload in demo
Date.now() returns the current time in milliseconds, but this isn't
strong enough to be a unique identifier, because two iterations can
happen inside the same millisecond.

We switch to generating random unique strings instead, as the nested
attributes feature of ActiveRecord and Sequel supports any keys, doesn't
have to be integers. We do it by combining Date.now() with Math.random().

Closes #290