Specify enctype="multipart/form-data" in multiple files guide
This is not necessary when doing direct uploads, because in that case
files are not actually uploaded through the form. But if people want to
do it without JavaScript, then they need this attribute. Rails adds it
automatically, but only when `#file_field` is called on the form object,
and in this case we're using a plain `#file_field_tag`, so we need to
add the `enctype="multipart/form-data"` tag manually.