Allow using custom signer for S3 object URLs
When wanting to serve private content through CloudFront, the S3 object
URLs need to be signed specially for CloudFront. To accommodate for
that, we add a :signer option to the S3 storage that takes a block,
which is called in #url with an unsigned URL and is expected return a
signed one.

  require "aws-sdk-cloudfront"

  signer = Aws::CloudFront::UrlSigner.new(
    key_pair_id:      "cf-keypair-id",
    private_key_path: "./cf_private_key.pem"
  )

  Shrine::Storage::S3.new(signer: signer.method(:signed_url))
  # or
  Shrine::Storage::S3.new(signer: -> (url, **options) { signer.signed_url(url, **options) })

See this article for more details:

https://docs.aws.amazon.com/AmazonCloudFront/latest/DeveloperGuide/PrivateContent.html