Encode filename for presign request in demo
fetch() doesn't properly URI-encode some characters (such as "%"), so
when Rack receives the request with query params that are not properly
encoded, it blows up.

So we use encodeURIComponent() to properly encode the filename. Since
fetch() won't re-encode characters that are already properly encoded,
this should now produce the correct behaviour.