Revert adding bucket name to URL path when :host and :force_path_style is used
This change was introduced in Shrine 2.11.0, but it causes a regression
when using an endpoint that requires :force_path_style (e.g. Minio), and
using a CDN which shouldn't have the bucket name included in the URL
path.

So we revert it, and we will instead make it possible to explicitly
specify host with a bucket name.