Fail when file command returned error in stdout
It appears that the file command can also fail with a successful exit
status, in which case the error is returned in stdout. This commit
handles that.

So far the only error we know about is

  cannot open: No such file or directory

It's very odd this error happens, because Shrine passes file content to
the standard input, it doesn't use a path on disk, so I don't know how
this error happen.

See https://github.com/shrinerb/shrine/issues/303