Force filename encoding to utf-8 in rack_file
Rack's multipart parser will return filename in binary encoding if it
has any non-ascii characters. When this file is converted to an IO-like
object via Shrine.rack_file and uploaded via Shrine#upload, the filename
will be copied to the Shrine::UploadedFile object.

However, when we attempt to serialize a Shrine::UploadedFile object with
a binary encoded filename into JSON, Ruby's JSON parser will raise an
exception because it requires all strings to be in UTF-8 encoding.

To address this, we force the filename encoding to UTF-8 in
Shrine.rack_file. This is similar to what Rails does in its
ActionDispatch::Http::UploadedFile.

Fixes #307