Update demo app for Uppy 0.28
Uppy 0.28 now passes cookies to presign requests to the Companion
endpoint, so now we can remove our custom implementation and let Uppy
make the request to our presign endpoint, which now sits on the same URL
path as the Companion endpoint. Since Shrine's presign endpoint is fully
compatible with Uppy Companion's endpoint, we can leave Uppy to think
it's making requests to a Companion app.

We also update the rest of the documentation to suggest "/s3/params",
so that it matches as closely to the walkthroughs as possible. The
presign_endpoint documentation itself remains as it is, because we won't
want to couple usage there in any way with Uppy.