Make S3#download and S3#open work with :sse_* options
Both S3#download and S3#open, in addition to the #get_object request,
are implicitly doing a #head_object request as well. S3#download does it
when retrieving Aws::S3::Object#content_type, while S3#open does it when
retrieving Aws::S3::Object#content_length.

When an S3 object is uploaded with :sse_* upload options (enabling
server side encryption), the same options need to be passed when
later downloading that object. Unfortunately, calling S3#download and
S3#open with :sse_* options is currently failing, because while the
options are successfully forwarded to the #get_object requests, they
were not forwarded to the #head_object requests, causing an
Aws::S3::Errors::BadRequest exception to be raised.

We fix this by eliminating the #head_object call in S3#download, while
in S3#open we make the #head_object explicit and pass any additional
options that are valid for #head_object.