Don't strip media type parameters in data_uri plugin
Currently `Shrine.data_uri(io)` is returning a DataFile object, where
the #content_type attribute holds the value of the media type from the
data URI, but with any media type parameters removed. This is incorrect
behaviour, as now the DataFile object isn't an accurate representation
of the data URI. If the user wants to retrieve the media type
parameters, they can't because that information is lost.

So we retain the media type with its parameters in
DataFile#content_type. The default Shrine#extract_mime_type method and
the :content_type analyzer from the determine_mime_type plugin will now
automatically strip away media type parameters, which is what we want.