Add :host option for UploadedFile#download_url
The download_endpoint already has a global :host option. However,
sometimes it's useful to be able to set the host per URL, for example if
we want to use multiple CDN hosts to get around browsers' limit of 6
concurrent requests per host, or if we want to infer the host from the
request context.

This commit adds the ability to pass :host to UploadedFile#download_url:

  uploaded_file.download_url(host: "https://example.com")