Make rack_response plugin work with Rack::Sendfile
The frontend server (e.g. NGINX) can be configured to serve static files,
which lifts the burden off of the web application. Rack applications can
utilize this with the Rack::Sendfile middleware. Rack::Sendfile will
detect when the response body object responds to `#to_path`, and given
that NGINX has been properly configured it will set the correct response
headers to signal NGINX to take over serving that file.

If the user is using FileSystem storage, they can benefit from this
optimization. So in this case we define `#to_path` on the response body
object. Now if the user happens to use Rack::Sendfile and has NGINX
properly configured, the file will automatically get handed off to
NGINX.