Refactor add_metadata plugin
This will make it easier to make the subsequent change that I want to
make. I also think it's generally nicer, as the `_metadata_proc` thing
was confusing to me, and I was the one who wrote it.

This makes a slight change in behaviour. Previously if a metadata block
would return nil, the metadata key wouldn't get added at all. Now the
metadata key does get added but with nil value. I think that's desired,
as we want to communicate that the metadata extraction has indeed run,
just that the return value was nil.

We also fix the overridden Shrine#extract_metadata now requiring the
context argument. The context argument should always be optional, so
that we can more easily call the method directly.