Don't mutate data in refresh_metadata plugin
When the data column is a JSON column, and we're using either Active
Record or Sequel ORMs, the column value will be returned as a Hash
(instead of a JSON string). This value is then passed directly to
Shrine::UploadedFile#initialize in Attacher#get.

In this case, if the refresh_metadata plugin mutates the data hash, it
mutates that column value that came from outside of Shrine. And it's
generally wrong to mutate objects that came from the outside, because
that can lead to hard to detect bugs.

Addresses https://github.com/shrinerb/shrine/issues/333