Fix backgrounding + refresh_metadata + JSON column not working
If the following was true:

* attachment data column is a JSON column,
* backgrounding and refresh_metadata plugins were loaded, and
* UploadedFile#refresh_metadata! was called during promotion,

then the promotion would abort.

The problem was that UploadedFile#refresh_metadata! was mutating the
data hash, which in the JSON column case was equal to the column value,
so when backgrounding plugin was checking whether the attachment
changed, it incorrectly concluded that it did because the metadata has
changed.

We modify the backgrounding plugin to use Shrine::UploadedFile#==
when checking whether the original attachment has changed. This method
only compared "id" and "storage" attributes, it doesn't look at
"metadata". That means it will return true if metadata happens to have
changed.

Addresses https://github.com/shrinerb/shrine/issues/333