Make urlsafe_serialization plugin private
The reason for extracting the urlsafe_serialization plugin from
download_endpoint was so that it can be reused in the upcoming
derivate_endpoint plugin. However, the derivate_endpoint plugin is
not ready yet and won't be released in the next Shrine version, so we
don't want to make the urlsafe_serialization plugin public yet, because
we might still need to modify it to support the needs of the
derivate_endpoint plugin. Once the derivate_endpoint plugin is released,
we'll release make urlsafe_serialiation plugin public as well.