Fix backgrounding plugin when storages are changed via Attachment.new
Until now the backgrounding plugin worked when Attachment.new set a
different storage, but the :cache and :store storages were still
registered. However, at the moment if :cache and :store didn't exist,
Attacher#swap was raising an error when reloading the record, because it
didn't use the Attacher set by the Attachment module. This commit fixes
that.

Closes #328, Closes #310