Drop support for MRI 2.1 and MRI 2.2
We've reached to a point where we need to add a runtime dependency that
only supports MRI 2.3 or higher. I don't know how to add a runtime
dependency conditionally based on the Ruby version. Normal conditionals
won't work, because the gemspec is evaluated when the gem is built.

Since MRI 2.1 and MRI 2.2 are EOL, I see no problem dropping support for
these Rubies.