Add tempfile plugin
The tempfile plugin makes it easier to reuse the same copy of the
uploaded file on disk during promotion. It adds UploadedFile#tempfile
method, which caches the downloaded content, and modifies
Shrine.with_file to call UploadedFile#tempfile when possible.

If we're extracting metadata during promotion and also doing
processing, we can use the tempfile plugin to ensure the uploaded file
is downloaded to disk only once during the whole process block.

  Shrine.plugin :tempfile

  class MyUploader < Shrine
    plugin :processing
    plugin :refresh_metadata
    plugin :add_metadata

    process(:store) do |io, context|
      io.open do
        io.refresh_metadata!

        original = io.tempfile
        # ... processing ...
      end
    end

    add_metadata(:foo) do |io, context|
      Shrine.with_file(io) { ... }
    end

    add_metadata(:bar) do |io, context|
      Shrine.with_file(io) { ... }
    end
  end

This is useful when dealing with large files, where every copy will be
expensive.