Yield new File on Shrine.with_file with tempfile plugin
If Shrine.with_file yields the UploadedFile#tempfile object directly,
then it's not thread-safe, because if multiple threads try to read the
same IO object at the same time, there will be race conditions.

We fix that by opening a new File object (with new file descriptor) to
the same path on disk, yielding that, and closing it after the block
finishes.