Fix ImageOptim example
ImageOptim#optimize_image is not safe, as it returns a path from a
Tempfile object, without keeping the reference to the Tempfile object.
This means that at any point Ruby's GC can delete the file on that path
before we're finished using it.

So we instead use the safer ImageOptim#optimize_image!, which doesn't
create new files.