Link to markdown docs on directly on GitHub
GitHub has a much nicer markdown formatting than the hanna-nouveau rdoc
formatter, so there is really no gain in linking to rdoc-processed
versions of these documents.

My only concern was that I didn't want the website to link to documents
on master branch, which might show functionality that hasn't yet been
released. But we easily solve that by scoping the revision of the latest
git tag, which equals to the last released version.

This also has the benefit of avoiding some bugs in rdoc's markdown
generator (which as far as I could see doesn't use any gem like
Kramdown), and allows us to use niceties of GFM such as tables. GitHub's
syntax highlighter is also much more robust, RDoc concludes pretty
easily that a code block is not Ruby code, so it often doesn't syntax
highlight.

We also avoid rdoc automatically linking every mention of the word
"Shrine" to the `Shrine` class rdoc, which was a bit confusing.

The fact that everything will be on GitHub also means we can more easily
use `#<reference>` anchors to link to different sections inside the same
document.

We also avoid README on master referencing guides that haven't
been rdoc-generated yet.

So yeah, we gain many benefits from this change.