Move :cache_control to Derivation option and simplify it
It's not necessary to allow users to override Cache-Control directives
individually using a Hash. We can just keep it simple and have them pass
the whole header value as a string. That way there is less magic and
it's obvious what the header value it is.

We also move :cache_control to a derivation option, which means it's
value can be set dynamically.