Return no "Content-Type" in #to_rack_response on unknown MIME type
When MIME type could not be determined, most MIME type analyzers and
Rack::Mime will return "application/octet-stream", which is just another
way of saying "this is a generic binary file".

However, just because *we* couldn't determine the MIME type, that
doesn't mean we shouldn't give the browser a try. Browsers can also do
some MIME type sniffing as well, which will take place if the
"Content-Type" response header is blank.

This solves a use case for derivation_endpoint where if :upload option
was enabled and derivation returned a WEBP image, the "Content-Type"
would be set to "application/octet-stream", because Rack::Mime doesn't
recognize the .webp extension. This would prevent the browser from
displaying the image inline, because even though the browser can figure
out that a WEBP image was returned, by setting "Content-Type" we're
telling the browser not to.