Don't return any "Content-Type" when unknown in derivation_endpoint
If "Content-Type" header value could not be determined from file
extension, were returning "application/octet-stream". However, this prevents the browser from
doing MIME type sniffing, and instead tells it that the response is a
generic binary file. This means, for example, that the browser will not
attempt to render the file inline. So, currently, if we're generating
WEBP images, since Rack::Mime doesn't recognize .webp yet, we'd prevent
the browser from determining that the file is a WEBP image.

So, instead of returning "application/octet-stream" when we failed to
determine the MIME type, we simply don't return any "Content-Type" in
that case. This allows browsers to determine the MIME type themselves.

While here, we also add some missing tests.