Revert "Extended determine MIME type with Marcel"
This reverts commit b1e5d8362a8204e287d1609b580ccb90f8f009d6.

The change was introduced in v2.15.0. Since the filename can be anything, it
was found in some small cases the mime-type determined based on filename can
be incorrect. This also created a potential security issue for an application.

Therefore the change is reverted and Shrine will have a method for users to
allow Marcel to fallback to using the filename if they choose to.

See issue #354.