Allow passing options to analzyers in determine_mime_type plugin
This commit allows passing in options to mime-type analyzers. For now, Marcel
is the only analyzer that makes use of any passed-in options.

Marcel analyzer takes one option called `:filename_fallback` which can be set
to true or false (default). When set to true, Marcel will also use the filename,
if available, to determine the mime-type when it fails to find it using the
magic number.