Fix versions being deleted when string names are used
When using strings for version names in the process block, after
promotion the uploaded files would get deleted. This is because when
Attacher#swap is called, it compares the hash of versions before and
after it has been persisted, and they will differ because the hash
loaded from the database will always have symbol keys.

Fixes #373