Add ETag header to prevent Rack::ETag from buffering file content
Rack::ETag is included by default in the Rails middleware stack, and it
will automatically generate digests from the response body for setting
the "ETag" header value. This will buffer the response body into memory,
causing higher memory usage and preventing potential streaming if the
webserver supports it (e.g. Falcon).

So we add an "ETag" header ourselves, calculated from the uploaded file
id. This causes Rack::ETag middleware to skip its thing. We use weak
validation and limit length to 32 as that's what Rack::ETag does.

To make the ETag as unique as possible, we calculate it from shrine
class, storage and id. That should ensure that only the same file on the
same storage has the same ETag.