Soft-rename upload_endpoint class to Shrine::UploadEndpoint
This makes the #inspect output of an instance of this class look
simpler, which is used in things like Rails' `rake routes` output. This
is also consistent with derivation_endpoint plugin's
Shrine::DerivationEndpoint class.

We deprecate the `Shrine::Plugins::UploadEndpoint::App` class in case
anyone was using it.

We also remove tests for the deprecated direct_upload plugin, because
the `Shrine::UploadEndpoint` constant in direct_upload plugin and
upload_endpoint are clashing.