Remove Roda dependency from download_endpoint
None of the other *_endpoint plugins depend on Roda anymore, so it's
time to do the same for download_endpoint plugin. This introduces
backwards incompatibility for anyone calling Roda methods on the
endpoint object, but that shouldn't be done, as Roda was just an
implementation detail.

The reason why this was done now and not in 3.0.0 is because in version
2.15.0 we changed `Shrine.download_method` to always return a new App
subclass. This meant the endpoint object was now an anonymous class,
because it wasn't assigned to a constant anymore. Unfortunately, that
doesn't work with Rails' `mount` method, as it expects classes to have a
name (i.e. to *not* be anonymous).