Remove backfilling `size` for uploads of unknown size from S3 storage
This complicates the code of the S3 storage, and that functionality
doesn't even belong there, it should be a separate plugin so that you
get it regardless of the storage.

This functionality wasn't documented nor it was requested, and it's
extremely rare that people will be uploading directly from
pipes/sockets, so I don't mind this minor backwards compatibility
breakage for the benefit of simpler code.