Add :url option to upload_endpoint
  plugin :upload_endpoint, url: true
  # or
  plugin :upload_endpoint, url: { public: true }
  # or
  plugin :upload_endpoint, url: -> (uploaded_file, request) {
    uploaded_file.url(**options)
  }

This option includes the uploaded file URL in the response body:

  {
    "data": { "id": "...", storage: "...", "metadata": {...} },
    "url": "https://example.com/path/to/file"
  }

I've seen people often using the :rack_response option to add the URL,
and this makes it more convenient.

This plays nicely with Uppy's xhr-upload plugin, since it automatically
detects the `url` field in the response body and assigns it to
`file.uploadURL`. That property is intended to be unified across
Uppy's destination plugins.