Add Shrine::Attachment() shorthand for Shrine::Attachment.new
Previously we had Shrine.attachment, but that doesn't really communicate
that the result is a module. So we add the Shrine.Attachment alias,
which can then idiomatically be called as Shrine::Attachment(). This
communicates that we're instantiating a Shrine::Attachment instance, and
it's a bit shorter than doing Shrine::Attachment.new.

I would like eventually convert all examples to use Shrine::Attachment(),
but we need to wait first until most people have upgraded.

This idea was taken from the dry-rb ecosystem, namely dry-equalizer.