Accept Uppy's default "files[]" param in upload_enpdoint
Previously we were only accepting a "file" param, which required us to
explicitly specify it when configuring Uppy's XHRUpload plugin:

  uppy.use(Uppy.XHRUpload, {
    endpoint: '/upload',
    fieldName: 'file',
  })

Now we also accept Uppy's default `files[]` array, assuming the array
will contain just a single file (that's Uppy's default behaviour), in
which case we can omit the `fieldName` option:

  uppy.use(Uppy.XHRUpload, { endpoint: '/upload' })

In the future we might support uploading multiple files in the same POST
request (Uppy's `bundle: true` option), but only if that feature is
requested.

We also refactor the code to make it easier to potentially add raw PUT
uploads in the near future.