Simplify default validation error messages
Parentheses split the message into two parts, and it's not needed, we
can say everything at once.

And it's not necessarily needed to hide the blacklist, that should be
the job of the user if they need to hide it. It's better to be explicit
by default and not assume too much, so we show the blacklist in the same
way as the whitelist.