Add shorthand aliases to metadata extraction plugins
This aliases

  Shrine.determine_mime_type(io)
  Shrine.extract_dimensions(io)
  Shrine.calculate_signature(io)

to

  Shrine.mime_type(io)
  Shrine.dimensions(io)
  Shrine.signature(io)

Which is much more convenient and therefore accessible. We still want to
keep the explicit versions, as they signal that these operations do some
work, so some users might prefer it.

This is just a small step toward moving metadata extraction from the
Shrine instance to the Shrine class.