Remove the "Settings" section from the README
It would rarely be obvious how to use it, as people are in different
contexts for different feature (Shrine or Attacher context). We can
explain this somewhere else in the future if needed, but I think it's
too low-level for the README, we want the README to stick to the
high-level concepts.