Add ability to skip and force metadata extraction
If metadata extraction is heavy, in tests people might want to skip it.
So we add support for that:

  uploaded_file = Shrine.upload(io, :store, metadata: false)
  uploaded_file.metadata #=> {}

When uploading UploadedFile objects, people might want to force metadata
extraction, e.g. when they want to extract metadata in background job
during promotion because they didn't do so in the controller for
performance reasons. So we add support for that as well:

  Shrine.upload(uploaded_file, :store, metadata: true)