Deprecate the copy plugin
The copy plugin has had some undesired behaviour, mostly around its
`#initialize_copy` override (see https://github.com/shrinerb/shrine/issues/237).

And it's simply not useful enough even for the `Attacher#copy` method.
The following:

  attacher.copy(other_attacher)

can be replaced with:

  copied_file = attacher.store.upload(other_attacher.get)
  attacher.set(copied_file)

In Shrine 3.0 this will be even simpler:

  attacher.attach(other_attacher.file)
  attacher.add_derivatives(other_attacher.deriatives) # if using versions

In conclusion, this plugin is not useful enough to be around.