Add a default logger
People should be able to have Shrine log anything into the Rails logger
when using Shrine inside a Rails app (same goes for any other web
framework). For example, in production it's important that any Shrine
warnings goes onto the same device as the rest of the application's
logs, for example a log file, so that the developer doesn't miss it.

  Shrine.logger = Rails.logger

This will also be needed for the upcoming instrumentation plugin.

We configure the default formatter to exclude timestamps and any other
information for cleaner logs (that's the default of Rails' logger
anyway). Users can always add this formatting if they want to.