Recommend using :marcel MIME type analyzer
While I decided that Shrine 3.0 will continue to not use an MIME type
analyzer by default, we can still recommend people to use the MIME type
analyzer we think it's best. In my opinion that's Marcel, because it's
used by Active Storage and it doesn't have any OS dependencies.

The default :file analyzer is nice in the sense that you don't have to
install any gems, but if you're deploying your app in a Docker container
which doesn't have the `file` command installed, MIME type extraction
suddenly won't work. So, while it will remain the default because you
don't need to add additional gems to the Gemfile, we'll still recommend
the :marcel analyzer in the documentation.