Inline extending/including modules in base
I like that when I require `lib/shrine/uploaded_file.rb`, that file is
in charge of providing me a `Shrine::UploadedFile` class with all
methods included. I don't think it should be the job for
`lib/shrine.rb`.

Sure, it adds a bit of duplication, but that's fine because the number
of core classes will not grow.