Deprecate the moving plugin
Users should now pass the `move: true` to FileSystem#upload. FileSystem
storage was the only one that supported moving anyway, so it didn't make
sense to have a whole plugin for that.

Users can replace the moving plugin with the following (provided that
both :cache and :store storages are FileSystem):

  plugin :upload_options, cache: { move: true }, store: { move: true }

Removing the moving plugin will simplify the Shrine class implementation
a bit, as well as the storage abstraction.