Deprecate the backup plugin
The backup plugin relies heavily on the Attacher implementation, which
will change in Shrine 3.0 in such a way that the behaviour of the backup
plugin will need to change as well, in a way which is significantly less
performant. So we deprecate the backup plugin.

The new preferred way of mirroring uploads is manually using the
instrumentation plugin:

https://github.com/shrinerb/shrine/wiki/Mirroring-Uploads