Add back support for MRI 2.3
Even though Ruby 2.3 is EOL, people still on Rails 3.2.x are stuck with
Ruby 2.3, because Ruby 2.4 has some backwards incompatible changes with
the JSON gem, and possibly other problems as well.

https://github.com/shrinerb/shrine/issues/365#issuecomment-513331477

Since it's really not a problem for us to maintain Ruby 2.3
compatibility (at least for now), we add back support for Ruby 2.3. This
will be backported into Shrine 2.19.1 as well.