Make :mini_mime the default extension inferrer
Capybara and Mail gem, both of which Rails depends on, switched from the
mime-types to the mini_mime gem, and it seems the rest of the Ruby
ecosystem is slowly moving to it as well.

So we make :mini_mime the default extension inferrer, since people are
more likely to have it in the Gemfile already compared to the mime-types
gem.