Replace source object metadata when uploading an S3 file
By default, when copying an S3 object, the destination object will
retain any source object metadata, and it's not possible to override it.
This has caused some confusion in the past with direct uploads, where
people were trying to override metadata when uploading cached file to
permanent storage, and that didn't work.

So we make it the default to tell S3 to drop any source object metadata
in the destination object, and allow setting any new metadata. This is a
safer default, because people expect to be able to override metadata.

It would be ideal if S3 would allow retaining any source object metadata
by default, but allow overriding it. But it doesn't, it goes
all-or-nothing, either you keep all source object metadata or none of
it.