Remove hooks plugin
It's complicating the internal implementation, and it's not worth it
anymore, now that we'll have much better low level API, as people should
have a much easier time porting their hooks into explicit code.

I know the hooks plugin hasn't been deprecated in 2.x, but that's
because the 2.x API wasn't good enough not to need hooks. But 3.x API
will be.