Remove parallelize plugin
It relies on the removed hooks plugin. With the new derivatives people
will be able to parallelize uploading/deleting versions themselves.

I originally wanted to just deprecate the parallelize plugin and remove
it in Shrine 4, but it's blocking simplification of the Shrine class, so
we remove it now. It's not an essential feature anyway.