Remove delete_promoted plugin
This deletion will no longer be backgroundable with new backgrounding
API, so it doesn't pay off anymore from the performance perspective.

However, now people will be able to disable caching. And if they're
using caching, they *should* be deleting files in a recurring script
anyway. This plugin is not useful enough anymore to continue to exist.