Add attacher_options plugin for passing options to attacher operations
There are many different attacher operations that could take options:

* attaching to cache
* attaching to store
* extracting metadata of cached file
* validating
* promoting
* destroying

The only way currently to pass options to these operations is via
Attacher#context. However, that term has become a bit of legacy, and it
doesn't really mean much.

I think it's very beneficial that users are able to specify options for
certain operations. They can currently do that if they're calling
operations directly, but with ORM plugins they're usually not, as these
operations happen automatically in callbacks.

So we add the attacher_options plugin which allows registering options
for specific operations, to be passed when the operation is called.

  attacher.attach_options(metadata: { "foo" => "bar" })
  attacher.attach(io)
  attacher.file.metadata["foo"] #=> "bar"

This also allows users to pass options to the backgrounding
promote/destroy blocks:

  Shrine::Attacher.promote_block do |attacher, **options|
    PromoteJob.perform_async(
      attacher.record,
      attacher.name,
      attacher.data,
      options[:current_user].id,
    )
  end

  # ...

  attacher.promote_options(current_user: current_user)
  attacher.promote_attached

It also allows disabling background promotion:

  attacher.promote_options(background: false)
  attacher.promote_attached # performs synchronously