Add Attacher#refresh_metadata! method
When using the model plugin, the user will usually want to have file
data with re-extracted metadata being written back to the attachment
attribute on the model instance. Now instead of doing it manually:

  attacher.file.refresh_metadata!
  attacher.set(attacher.file)

the user can call #refresh_metadata! on the attacher which will do it
automatically:

  attacher.refresh_metadata! # writes file data back into the attachment attribute

This is very convenient when re-extracting metadata in a background job,
as it can be nicely coupled with #atomic_persist:

  class ExtractMetadataJob
    def perform(record, name, data)
      attacher = Shrine::Attacher.retrieve(model: record, name: name, data: data)
      attacher.refresh_metadata!
      attacher.atomic_persist
    end
  end