Use load/dump interface for column serializer
YAML, Oj, MultiJson, MessagePack and probably many other libraries
implement the load/dump interface, so it's more convenient to support
that interface instead of generate/parse, as that way people don't need
to create custom objects, they can just pass the library constant
directly.

  plugin :column, serializer: Oj

The only issue was that for the default serializer, the JSON standard
library, we have to use JSON.generate/JSON.parse because
JSON.load/JSON.dump should not be used with untrusted input. But we work
around that by creating a wrapper class that implements the load/dump
interface.