Default to permanent storage in Attacher#upload
The main storage in the Attacher is the permanent storage, the temporary
storage is just there to help with the attachment flow when user uploads
a file via a form. It's also possible to use the attacher without a
temporary storage, so we want to make it as easy as possible to work
with the permanent storage. So we make Attacher#upload use permanent
storage by default if no storage was specified.

This also simplifies some tests in which we don't care to which storage
a file was uploaded to.