Deep duplicate Shrine.opts on subclassing
Previously Shrine.opts was duplicated only one level in depth, which
means that plugins needed to put hashes/arrays that were being modified
into top level, to ensure they're duplicated on the subclass. This is
lame and causes additional mental overhead.

So, we deep duplicate Shrine.opts and move some collections to the top
level.