Allow backgrounding block which evaluates inside Attacher instance
We want Shrine to be as convenient as possible to use. This can be a lot
to write:

  Shrine::Attacher.promote_block do |attacher|
    PromoteJob.perform_async(
      attacher.record.class,
      attacher.record.id,
      attacher.name,
      attacher.data,
    )
  end

We can make it easier by evaluating the block inside an attacher
instance when block doesn't have arguments:

  Shrine::Attacher.promote_block do
    PromoteJob.perform_async(record.class, record.id, name, data)
  end

The important thing is that we allow both block styles, so that people
can choose to be explicit or implicit.