Rename default_url_options plugin to url_options
People have often missed that this plugin exists, and I think it's
partly because it's one of few plugins that start with "default_*".
Personally I always disliked how long the name was, and that it's
inconsistent with the upload_options plugin which does basically the
same thing just for upload options.

So we rename the default_url_options plugin to just url_options, and
keep the deprecated default_url_options alias.