Add Shrine::Storage::Memory from shrine-memory gem
Shrine::Storage::Memory is something I use a lot, especially in
self-contained scripts where I don't want any side effects, and I always
kind of minded that I have to pull it from a separate gem.

It *is* meant as a mock storage and not a real storage, but why not make
mocking more convenient in Shrine by shipping Shrine::Storage::Memory
with it? I think the convenience pays off.