Require same uploader class in UploadedFile#==
Currently UploadedFile#== allows the uploader class of the file on the
right hand side to be a subclass of the uploader class of the file on
the left hand side. However, that means the #== operator isn't
symmetrical, this rule would need to be applied regardless of the order
of operands.

We don't need that, however, we can just require that the classes match
exactly. It's very unlikely that anyone was using this behaviour.