Fix Attacher.retrieve not working when called on Shrine::Attacher
When Shrine::Attachment module is used on a model/entity,
Attacher.retrieve would fail with Shrine::AttachmentChanged, because
when it was comparing files it was instantiating UploadedFile using the
receiver attacher class (which when following the docs is
Shrine::Attacher). This caused UploadedFile#== to return the false,
because even though the two UploadedFile objects had the same data, they
were of different classes.

We fix that by instantiating the attacher class of the model/entity
attachment.