Raise Shrine::FileNotFound in Storage#open
In download_endpoint and derivation_endpoint it would be really useful
to know when Storage#open failed because file was not found. Currently
we're making a Storage#exists? call to check whether the file exists,
which adds unnecessary overhead. Moreover, in derivation_endpoint the
user currently needs to specify which exceptions should be translated to
a 404.

So we make storages raise Shrine::FileNotFound in Storage#open, and make
this a convention that all storages should follow.