Evaluate derivation block in context of a Shrine::Derivation instance
It is not useful at all to evaluate it in context of an uploader
instance, because this doesn't give us any useful information. In
general we want to be slowly moving away from an uploader instance and
instead rely more on class-level API.

Moreover, in the near future we will want to allow generating
Shrine::Derivation objects with custom source file that is not a
Shrine::UploadedFile object, so in this case we wouldn't even have
access to the uploader instance.

So, we instead evaluate the derivation block inside a context of the
Shrine::Derivation instance. This allows users to access options for
that derivation, which means they can use this information during
processing.