Remove extra Storage#exists? calls in derivation_endpoint
When :upload is set but no :upload_redirect, we're always first checking
whether the derivation result exists on the storage, and only then
streaming it to the response body. The extra Storage#exists? call is
adding overhead to the request, which might be significant depending on
the circumstances.

Now that Storage#open raises Shrine::FileNotFound, we can avoid the
extra Storage#exists? call. We'll see how that affects performance.

As a byproduct, a new Shrine::Derivation#opened method was added for
returning an opened Shrine::UploadedFile object pointing to the
derivation result if it exists.

We also stop handling derivative file being deleted on upload, which
could be the case when delete_raw plugin or `move: true` is used for
FileSystem storage. It was difficult to the original change and still
keep support for this, so I decided it was not worth it. People should
not configure their uploaders to automatically delete uploaded files.