Eliminate #head_object request in S3#open
We want to return object's content length in S3#open, so that
UploadedFile#size can retrieve it when metadata is being refreshed.
But, because aws-sdk-s3 doesn't expose this when using
Aws::S3::Object#get or Aws::S3::Client#get_object directly, we hacked
our way around it by making a separate #head_object request to retrieve
the content length.

However, that comes with a performance penalty, especially for small
files where the HEAD object takes around the same time as the GET object
would take. Also, it wasn't great how we needed to mangle params to
ensure we're passing SSE and such to the #head_object request.

So we come up with a different solution, one which does technically rely
on private S3 API, but IMO is unlikely to change. We build the request,
start it, then retrieve content length from headers that S3 writes
before it starts streaming the body. I'm satisfied with this solution.

As a side effect, we remove the Aws::S3::Object from Down::Chunked#data,
because it won't have any data anymore, and trying to access data will
trigger a #head_object request.